Split up application.js
